Improved error handling: parser+printer results
This commit includes an application wide refactoring that wraps the result of
every parser and printer in a new type, ParserResult and PrinterResult,
respectively, that encapsulates the result and parsing/printing and any
warnings/errors encountered along the way.

Furthermore, much of the code has been streamlined and draft implementations for
utility functions for error reporting has been implemented.