Contexts: restructuring -> more encapsulation
Trying to restructure the project a bit such that everything parser related is
found inside the `lib/parser` folder and that `lib/parser.ex` defines the
'facade' of the these submodules. Likewise, for printer related modules in
`lib/printer`.  However, this exercise isn't very doable for the `lib/types`
folder, so have to consider if I'm gaining anything from going down this road
for types.