Contexts (#58)
* Contexts: restructuring -> more encapsulation

Trying to restructure the project a bit such that everything parser related is
found inside the `lib/parser` folder and that `lib/parser.ex` defines the
'facade' of the these submodules. Likewise, for printer related modules in
`lib/printer`.  However, this exercise isn't very doable for the `lib/types`
folder, so have to consider if I'm gaining anything from going down this road
for types.

* Adds contexts 'parser' and 'printer'

We isolate all parser related code in the 'parser' folder and expose the
parser.ex module as the interface for that folder / context.

We repeat this exercise for the 'printer' folder / context.

Furthermore, we remove all the import statements and replace them with more
explicit aliases.

Closes #57