Adds 'print_fuzzer' function to printers
Each printer now has a 'print_fuzzer' which prints a fuzzer for that type along
with a fuzzing test that checks that encoding followed by decoding of that type
results in the original type.

This commit has refactors all printer tests and adjusts code according to
updated credo and dialyzer warnings.