Updates json_schema dependency to 0.2.0
Given that `json_schema` has breaking changes in the way it handles references,
such that both local and global references are now handled with the `URI` type,
this has required minor changes in all of the printer modules, and as a result
I've decided to bump the minor version of json-schema-to-elm given the scope of
affected files. However, this change should make things easier going forward
when implementing new feature.